rename module to `on_disk_point_to_values`
