+++ lib/segment/src/index/field_index/geo_index/mmap_geo_index/mod.rs
@@ -6,7 +6,7 @@ use common::universal_io::{MmapFile, TypedStorage, UniversalRead};
 use serde::{Deserialize, Serialize};
 
 use crate::index::field_index::geo_hash::GeoHashRaw;
-use crate::index::field_index::stored_point_to_values::OnDiskPointToValues;
+use crate::index::field_index::on_disk_point_to_values::OnDiskPointToValues;
 use crate::types::GeoPoint;
 
 mod lifecycle;
