+++ lib/segment/src/index/field_index/map_index/on_disk_map_index/lifecycle.rs
@@ -21,7 +21,7 @@ use super::{
 };
 use crate::common::Flusher;
 use crate::common::operation_error::{OperationError, OperationResult};
-use crate::index::field_index::stored_point_to_values::OnDiskPointToValues;
+use crate::index::field_index::on_disk_point_to_values::OnDiskPointToValues;
 
 impl<N, S> OnDiskMapIndex<N, S>
 where
