+++ lib/segment/src/index/field_index/numeric_index/immutable_numeric_index/lifecycle.rs
@@ -14,7 +14,7 @@ use crate::common::operation_error::OperationResult;
 use crate::index::field_index::histogram::Histogram;
 use crate::index::field_index::immutable_point_to_values::ImmutablePointToValues;
 use crate::index::field_index::numeric_point::{Numericable, Point};
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 
 impl<T, S> ImmutableNumericIndex<T, S>
 where
