+++ lib/segment/src/index/field_index/numeric_index/immutable_numeric_index/mod.rs
@@ -8,7 +8,7 @@ use super::on_disk_numeric_index::OnDiskNumericIndex;
 use crate::index::field_index::histogram::Histogram;
 use crate::index::field_index::immutable_point_to_values::ImmutablePointToValues;
 use crate::index::field_index::numeric_point::{Numericable, Point};
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 
 mod lifecycle;
 mod read_ops;
