+++ lib/segment/src/index/field_index/numeric_index/on_disk_numeric_index/mod.rs
@@ -6,7 +6,7 @@ use common::universal_io::{MmapFile, TypedStorage, UniversalRead};
 use super::Encodable;
 use crate::index::field_index::histogram::Histogram;
 use crate::index::field_index::numeric_point::{Numericable, Point};
-use crate::index::field_index::stored_point_to_values::{OnDiskPointToValues, StoredValue};
+use crate::index::field_index::on_disk_point_to_values::{OnDiskPointToValues, StoredValue};
 
 mod lifecycle;
 mod read_ops;
