+++ lib/segment/src/index/field_index/numeric_index/on_disk_numeric_index/read_ops.rs
@@ -16,7 +16,7 @@ use super::OnDiskNumericIndex;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::histogram::Histogram;
 use crate::index::field_index::numeric_point::{Numericable, Point};
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 use crate::index::payload_config::StorageType;
 
 impl<T: Encodable + Numericable + Default + StoredValue + 'static, S: UniversalRead>
