+++ lib/segment/src/index/field_index/numeric_index/query.rs
@@ -23,8 +23,8 @@ use super::Encodable;
 use super::numeric_index_read::NumericIndexRead;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::numeric_point::{Numericable, Point};
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 use crate::index::field_index::stat_tools::estimate_multi_value_selection_cardinality;
-use crate::index::field_index::stored_point_to_values::StoredValue;
 use crate::index::field_index::utils::check_boundaries;
 use crate::index::field_index::{CardinalityEstimation, PayloadBlockCondition, PrimaryCondition};
 use crate::index::query_optimization::optimized_filter::ConditionCheckerFn;
