+++ lib/segment/src/index/field_index/numeric_index/read_only/lifecycle.rs
@@ -10,7 +10,7 @@ use super::super::storage::read_only::ReadOnlyNumericIndexInner;
 use super::ReadOnlyNumericIndex;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::numeric_point::Numericable;
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 use crate::index::payload_config::IndexMutability;
 
 impl<T: Encodable + Numericable + StoredValue + Send + Sync + Default, P, S: UniversalRead>
