+++ lib/segment/src/index/field_index/numeric_index/read_only/value_retriever.rs
@@ -18,7 +18,7 @@ use super::super::numeric_index_read::NumericIndexRead;
 use super::ReadOnlyNumericIndex;
 use crate::common::utils::MultiValue;
 use crate::index::field_index::numeric_point::Numericable;
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 use crate::index::query_optimization::rescore_formula::value_retriever::VariableRetrieverFn;
 use crate::types::{
     DateTimePayloadType, FloatPayloadType, IntPayloadType, UuidIntType, UuidPayloadType,
