+++ lib/segment/src/index/field_index/numeric_index/storage/mod.rs
@@ -33,7 +33,7 @@ use super::immutable_numeric_index::ImmutableNumericIndex;
 use super::mutable_numeric_index::MutableNumericIndex;
 use super::on_disk_numeric_index::OnDiskNumericIndex;
 use crate::index::field_index::numeric_point::Numericable;
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 
 pub enum NumericIndexInner<T: Encodable + Numericable + StoredValue + Send + Sync + Default>
 where
