+++ lib/segment/src/index/field_index/numeric_index/storage/read_only/trait_impls.rs
@@ -15,7 +15,7 @@ use super::super::super::{Encodable, query};
 use super::ReadOnlyNumericIndexInner;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::numeric_point::Numericable;
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 use crate::index::field_index::{
     CardinalityEstimation, PayloadBlockCondition, PayloadFieldIndexRead,
 };
