+++ lib/segment/src/index/field_index/numeric_index/tests.rs
@@ -16,7 +16,7 @@ use tempfile::{Builder, TempDir};
 use super::*;
 use crate::common::operation_error::OperationResult;
 use crate::index::field_index::numeric_point::Numericable;
-use crate::index::field_index::stored_point_to_values::StoredValue;
+use crate::index::field_index::on_disk_point_to_values::StoredValue;
 use crate::index::field_index::{
     CardinalityEstimation, FieldIndexBuilderTrait, PayloadFieldIndexRead, ValueIndexer,
 };
